requirements.txt
django==2.1.0
django-extensions
jsonfield
pillow
requests
beautifulsoup4
jupyter
python-telegram-bot

pandas
lxml
